backend/requirements.txt
fastapi
uvicorn
python-docx
openai
python-dotenv
pydantic
python-multipart
jinja2
gunicorn
uvloop
httptools
